requirements.txt
bleach
boto3
bs4
django>=3.1
djangorestframework
elasticsearch
elasticsearch-dsl
typing
uwsgi
warcio
